ml-model/requirements.txt
# Core ML Dependencies
torch>=1.12.0
torchvision>=0.13.0

# YOLOv5 Dependencies
opencv-python>=4.6.0
matplotlib>=3.5.0
seaborn>=0.11.0

# AWS and Deployment
sagemaker>=2.150.0
boto3>=1.26.0

# Data Processing
pillow>=9.0.0
pandas>=1.4.0
pyyaml>=6.0
numpy>=1.21.0

# Additional utilities
tqdm>=4.64.0
requests>=2.28.0
